radius-client: shorten timeout for accounting requests
Accounting requests are blocking in the current implementation.
Failure from the remote RADIUS accounting server to reply in a timely
fashion may unnecessarily lengthen the delay before which the client is
granted access, *even if* the RADIUS _authentication_ server did grant
such access, especially as the timeout is compounded with the number of
retries ("1", for a total of 10s).

Therefore, apply a shorter timeout to RADIUS accounting requests which
are not strictly necessary for the system to operate correctly.

Signed-off-by: Thibaut VARÈNE <hacks@slashdirt.org>