uspot: client_get(): kick client when timeout is exceeded
The uspot housekeeping loop runs at 10s interval. When a session timeout
is set, the timeout can fall between the loop runs, resulting in
negative values being reported to the API and clients being prompted for
reconnection only to be told "you are connected".

Fix this by asynchronously kicking clients in the client_get() routine
when the timeout is exceeded.

Signed-off-by: Thibaut VARÈNE <hacks@slashdirt.org>