uspotfilter: rework neigh management
Linux may aggressively delete neighbours that are not necessarily FAILED,
in particular it may delete some STALE neighs, only to reach them again
later. This is particularly true with the default gc_neighN values which
are fairly low for a busy router.

This commit works around this by deleting clients when they have been
marked as FAILED or their FAILED neigh has been deleted. All other
deleted neigh states will result in marking the client idle if it was
allowed (which will then be purged by uspot as needed) or deleting if not.

Signed-off-by: Thibaut VARÈNE <hacks@slashdirt.org>