uspot: use a single operation for client removal
When a client is terminated, whatever the reason may be, they must be
removed from the spotfilter backend just like they are removed from the
uspot one, there is no reason to treat some removal causes differently.

Furthermore, this causes lingering data in uspotfilter, and eventually
leads to a runaway situation as state becomes desynchronised between
uspot and uspotfilter.

Fix this by using a single operation for all removal cases.

Signed-off-by: Thibaut VARÈNE <hacks@slashdirt.org>