uspot: implement state_get()/state_set()
This implements a basic way to dynamically and gracefully activate or
deactivate uspot. When deactivated, uspot will prompt clients with an
explicit "service is currently unavailable" message.

This is useful in conjunction with e.g. a WAN watchdog that can thus
update uspot state when WAN connection is lost so that users get a more
user-friendly message.

Signed-off-by: Thibaut VARÈNE <hacks@slashdirt.org>